Mark

+++ app/components/CrapIdLayout/CrapIdLayout.js
@@ -1,14 +1,23 @@
 "use client";
 
-import { useRouter } from "next/navigation";
 import styles from "./CrapIdLayout.module.css";
 import Image from "next/image";
 import { getSessions } from "@/app/actions";
+import { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
 
-const CrapIdLayout = ({ data }) => {
-  const crap = data.data;
-  console.log(crap);
+const CrapIdLayout = ({ data, id }) => {
   const router = useRouter();
+  const crap = data.data;
+  const [isOwner, setIsOwner] = useState(true);
+
+  useEffect(() => {
+    if (crap.owner._id !== id) {
+      return;
+    } else {
+      setIsOwner(false);
+    }
+  }, [id, crap.owner._id]);
 
   const deleteCrap = async () => {
     const base =
@@ -18,10 +27,35 @@ const CrapIdLayout = ({ data }) => {
 
     const token = await getSessions();
 
+    if (crap.owner._id !== id) {
+      return;
+    }
+
+    await fetch(`${base}api/crapId?token=${token?.value}&id=${crap._id}`, {
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token?.value}`,
+      },
+    });
+
+    if (crap.owner._id !== id) {
+      return;
+    }
+    return router.push("/mine");
+  };
+
+  const interested = async () => {
+    const base =
+      process.env.NODE_ENV === "development"
+        ? "http://localhost:3000/"
+        : "https://madd9022-finalproject.vercel.app/";
+
+    const token = await getSessions();
+
     const resp = await fetch(
-      `${base}api/delete?token=${token?.value}&id=${crap._id}`,
+      `${base}api/crapId?token=${token?.value}&id=${crap._id}`,
       {
-        method: "DELETE",
+        method: "POST",
         headers: {
           Authorization: `Bearer ${token?.value}`,
         },
@@ -30,47 +64,82 @@ const CrapIdLayout = ({ data }) => {
 
     const data = await resp.json();
 
-    return router.push("/mine");
+    router.refresh();
   };
 
   return (
     <div>
       <div className={styles.container}>
-        <form className={styles.form}>
-          <div className={styles.formBox}>
-            <label className={styles.label} htmlFor="address">
-              Pickup Address:
-            </label>
-            <input
-              className={styles.inputField}
-              type="text"
-              placeholder="Address"
-              name="address"
-            />
-          </div>
-          <div className={styles.formBox}>
-            <label className={styles.label} htmlFor="title">
-              Pickup Date:
-            </label>
-            <input className={styles.inputField} type="date" name="date" />
-          </div>
-          <div className={styles.formBox}>
-            <label className={styles.label} htmlFor="title">
-              Pickup Time:
-            </label>
-            <input className={styles.inputField} type="time" name="time" />
-          </div>
+        {isOwner ? (
+          <>
+            {crap.status === "AVAILABLE" ? (
+              <div>
+                <p className={styles.interestText}>
+                  Are you interested in this crap? Click the button below
+                </p>
+                <button className={styles.interestBtn} onClick={interested}>
+                  INTERESTED
+                </button>
+              </div>
+            ) : (
+              <div>
+                <p className={styles.interestText}>
+                  Waiting for seller to respond
+                </p>
+              </div>
+            )}
+          </>
+        ) : (
+          <form className={styles.form}>
+            <div className={styles.formBox}>
+              <label className={styles.label} htmlFor="address">
+                Pickup Address:
+              </label>
+              <input
+                className={styles.inputField}
+                type="text"
+                placeholder="Address"
+                name="address"
+              />
+            </div>
+            <div className={styles.formBox}>
+              <label className={styles.label} htmlFor="title">
+                Pickup Date:
+              </label>
+              <input className={styles.inputField} type="date" name="date" />
+            </div>
+            <div className={styles.formBox}>
+              <label className={styles.label} htmlFor="title">
+                Pickup Time:
+              </label>
+              <input className={styles.inputField} type="time" name="time" />
+            </div>
 
-          <div className={styles.formBox}>
-            <button className={styles.submit} type="submit">
-              Suggest pickup time and location
-            </button>
-          </div>
-        </form>
+            <div className={styles.formBox}>
+              <button className={styles.submit} type="submit">
+                Suggest pickup time and location
+              </button>
+            </div>
+          </form>
+        )}
+
+        {!isOwner ? (
+          <>
+            {crap.status === "AVAILABLE" ? (
+              <div>
+                <p>No one has shown interested in this crap</p>
+              </div>
+            ) : (
+              ""
+            )}
+          </>
+        ) : (
+          ""
+        )}
 
         <div className={styles.card}>
           <div className={styles.cardTitle}>
-            {crap.title} - {crap.status}
+            {crap.title} - [{crap.status}]
           </div>
 
           <div className={styles.cardContent}>{crap.description}</div>
@@ -86,11 +155,15 @@ const CrapIdLayout = ({ data }) => {
             />
           </div>
 
-          <div>
-            <button className={styles.delete} onClick={deleteCrap}>
-              Delete this crap
-            </button>
-          </div>
+          {isOwner ? (
+            "Only the owner can delete this crap."
+          ) : (
+            <div>
+              <button className={styles.delete} onClick={deleteCrap}>
+                Delete this crap
+              </button>
+            </div>
+          )}
         </div>
       </div>
     </div>
